+++ src/main/java/com/evacipated/cardcrawl/modthespire/Loader.java
@@ -7,6 +7,10 @@ import com.evacipated.cardcrawl.modthespire.ui.Launcher;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonSyntaxException;
+import com.evacipated.cardcrawl.modthespire.ui.ModSelectWindow;
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import com.google.gson.JsonSyntaxException;
 import com.google.gson.reflect.TypeToken;
 import com.vdurmont.semver4j.Semver;
 import javassist.ClassPool;
@@ -28,6 +32,7 @@ import java.nio.file.Paths;
 import java.util.List;
 import java.util.Timer;
 import java.util.*;
+import java.util.function.BiConsumer;
 
 public class Loader
 {
@@ -36,6 +41,7 @@ public class Loader
 
     public static Semver MTS_VERSION;
     public static String MOD_DIR = "mods/";
+    private static String BETA_SUBDIR = "beta/";
     public static String STS_JAR = "desktop-1.0.jar";
     private static String MAC_STS_JAR = "SlayTheSpire.app/Contents/Resources/" + STS_JAR;
     private static String STS_JAR2 = "SlayTheSpire.jar";
@@ -268,14 +274,35 @@ public class Loader
             e.printStackTrace();
         }
 
-//        findGameVersion();
 
+        // Save workshop locations
+        if (!workshopInfos.isEmpty()) {
+            try {
+                List<String> workshopLocations = new ArrayList<>();
+                for (SteamSearch.WorkshopInfo info : workshopInfos) {
+                    if (info == null) {
+                        continue;
+                    }
+                    workshopLocations.add(info.getInstallPath().toAbsolutePath().toString());
+                }
+
+                Gson gson = new GsonBuilder().setPrettyPrinting().create();
+                String data = gson.toJson(workshopLocations);
+                Files.write(Paths.get(SpireConfig.makeFilePath(null, "WorkshopLocations", "json")), data.getBytes());
+            } catch (IOException e) {
+                // TODO
+                e.printStackTrace();
+            }
+        }
+
+//        findGameVersion();
+//
 //        EventQueue.invokeLater(() -> {
 //            ModInfo[] modInfos = getAllMods(workshopInfos);
-//            ex = new Launcher(modInfos);
+//            ex = new ModSelectWindow(modInfos);
 //            ex.setVisible(true);
 //
-////            ex.warnAboutMissingVersions();
+//            ex.warnAboutMissingVersions();
 //
 //            String java_version = System.getProperty("java.version");
 //            if (!java_version.startsWith("1.8")) {
@@ -283,7 +310,7 @@ public class Loader
 //                JOptionPane.showMessageDialog(null, msg, "Warning", JOptionPane.WARNING_MESSAGE);
 //            }
 //
-////            ex.startCheckingForMTSUpdate();
+//            ex.startCheckingForMTSUpdate();
 //        });
     }
 
@@ -503,6 +530,17 @@ public class Loader
     {
         List<ModInfo> modInfos = new ArrayList<>();
 
+        // Beta version of mods
+        if (STS_BETA) {
+            for (File f : getAllModFiles(MOD_DIR + BETA_SUBDIR)) {
+                ModInfo info = ModInfo.ReadModInfo(f);
+                if (info != null) {
+                    if (modInfos.stream().noneMatch(i -> i.ID == null || i.ID.equals(info.ID))) {
+                        modInfos.add(info);
+                    }
+                }
+            }
+        }
         // "mods/" directory
         for (File f : getAllModFiles(MOD_DIR)) {
             ModInfo info = ModInfo.ReadModInfo(f);
@@ -513,37 +551,84 @@ public class Loader
             }
         }
 
-        // Workshop content
-        for (SteamSearch.WorkshopInfo workshopInfo : workshopInfos) {
-            for (File f : getAllModFiles(workshopInfo.getInstallPath().toString())) {
-                ModInfo info = ModInfo.ReadModInfo(f);
-                if (info != null) {
-                    // Disable the update json url for workshop content
-                    info.UpdateJSON = null;
-                    info.isWorkshop = true;
+        BiConsumer<File, Boolean> lambda = (f, beta) -> {
+            ModInfo info = ModInfo.ReadModInfo(f);
+            if (info != null) {
+                // Disable the update json url for workshop content
+                info.UpdateJSON = null;
+                info.isWorkshop = true;
 
-                    // If the workshop item is a newer version, use it instead of the local mod
-                    boolean doAdd = true;
-                    Iterator<ModInfo> it = modInfos.iterator();
-                    while (it.hasNext()) {
-                        ModInfo modInfo = it.next();
-                        if (modInfo.ID != null && modInfo.ID.equals(info.ID)) {
-                            if (modInfo.ModVersion == null || info.ModVersion == null) {
-                                doAdd = false;
-                                break;
-                            }
-                            if (info.ModVersion.isGreaterThan(modInfo.ModVersion)) {
-                                it.remove();
-                            } else {
-                                doAdd = false;
-                                break;
-                            }
+                // If the workshop item is a newer version, use it instead of the local mod
+                boolean doAdd = true;
+                Iterator<ModInfo> it = modInfos.iterator();
+                while (it.hasNext()) {
+                    ModInfo modInfo = it.next();
+                    if (modInfo.ID != null && modInfo.ID.equals(info.ID)) {
+                        if (modInfo.ModVersion == null || info.ModVersion == null) {
+                            doAdd = false;
+                            break;
+                        }
+                        if (info.ModVersion.isGreaterThan(modInfo.ModVersion)) {
+                            it.remove();
+                        } else {
+                            doAdd = false;
+                            break;
                         }
                     }
-                    if (doAdd) {
-                        modInfos.add(info);
+                }
+                if (doAdd) {
+                    modInfos.add(info);
+                }
+            }
+        };
+        // Workshop content
+        if (!workshopInfos.isEmpty()) {
+            for (SteamSearch.WorkshopInfo workshopInfo : workshopInfos) {
+                // Normal
+                for (File f : getAllModFiles(workshopInfo.getInstallPath().toString())) {
+                    lambda.accept(f, false);
+                }
+                // Beta
+                if (STS_BETA) {
+                    for (File f : getAllModFiles(Paths.get(workshopInfo.getInstallPath().toString(), BETA_SUBDIR).toString())) {
+                        lambda.accept(f, true);
                     }
                 }
+            }
+        } else {
+            // Load workshop locations
+            try {
+                List<String> workshopLocations = null;
+                String path = SpireConfig.makeFilePath(null, "WorkshopLocations", "json");
+                if (new File(path).isFile()) {
+                    String data = new String(Files.readAllBytes(Paths.get(path)));
+                    Gson gson = new Gson();
+                    Type type = new TypeToken<List<String>>(){}.getType();
+                    try {
+                        workshopLocations = gson.fromJson(data, type);
+                    } catch (JsonSyntaxException ignore) {
+                        workshopLocations = null;
+                    }
+                }
+                if (workshopLocations == null) {
+                    workshopLocations = new ArrayList<>();
+                }
+
+                for (String location : workshopLocations) {
+                    // Normal
+                    for (File f : getAllModFiles(location)) {
+                        lambda.accept(f, false);
+                    }
+                    // Beta
+                    if (STS_BETA) {
+                        for (File f : getAllModFiles(Paths.get(location, BETA_SUBDIR).toString())) {
+                            lambda.accept(f, true);
+                        }
+                    }
+                }
+            } catch (IOException e) {
+                // TODO
+                e.printStackTrace();
             }
         }